Move functions to different modules per code review recommendation
